Removes ModuloEventSequenceEncoderDecoder, and moves ModuloPerformanceEventSequenceEncoderDecoder to performance_encoder_decoder.py.
